refactor: update Header import to use the main Header component across multiple pages
